Require recent pip version
Versions >= 7 of pip have caching of binary wheels. This gives
quite a dramatic speedup in creating new virtual environments
and hence in total deployment.

Unfortunately, the Ansible pip module doesn't allow us to just
specify a minimal version, only an exact version:

https://github.com/ansible/ansible-modules-core/blob/22bfb54d9d8ba784425d64bd4094b0153b159806/packaging/language/pip.py#L164

The current pip release is version 7.1.2, so we fix it at that.